+++ dart/lib/src/tag/tag.dart
@@ -7,8 +7,10 @@ library;
 
 import 'package:flutter/widgets.dart' hide Action;
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:layrz_logging/layrz_logging.dart';
 import 'package:layrz_sdk/src/access/access.dart';
 import 'package:layrz_sdk/src/actions/actions.dart';
+import 'package:layrz_sdk/src/api/api.dart';
 import 'package:layrz_sdk/src/asset/asset.dart';
 import 'package:layrz_sdk/src/avatar/avatar.dart';
 import 'package:layrz_sdk/src/care_protocols/care_protocols.dart';